--- Domain/Sources/Errors/Languages/Repositories/SetLanguagesRepositoryError.swift
@@ -1,21 +1,0 @@
-import Foundation
-
-public enum SetLanguagesRepositoryError: LocalizedError, Sendable {
-    /// 작업 취소의 경우
-    case cancelled
-    /// 언어 설정 저장에 실패한 경우
-    case saveFailed
-    /// 알 수 없는 Error
-    case unknown(Error)
-
-    public var errorDescription: String? {
-        switch self {
-        case .cancelled:
-            return nil
-        case .saveFailed:
-            return "언어 설정 저장에 실패했습니다"
-        case .unknown(let error):
-            return error.localizedDescription
-        }
-    }
-}
