+++ Domain/Sources/Interfaces/VoiceNotes/STTRepository.swift
